backend/pyproject.toml
[project]
name = "website-ingestion-rag"
version = "0.1.0"
description = "Website Ingestion & Vectorization for RAG"
authors = [{name = "Claude Code", email = "claude@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "cohere>=4.0.0",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"